+++ __tests/wow_api/API/Enums.lua
@@ -107,4 +107,12 @@ Enum = {
 		Fresh = 11,
 		FreshHardcore = 12,
 	},
+	LootMethod = {
+		Freeforall = 0,
+		Roundrobin = 1,
+		Masterlooter = 2,
+		Group = 3,
+		Needbeforegreed = 4,
+		Personal = 5,
+	},
 }