+++ changelog.md
@@ -1,3 +1,8 @@
+## Bugfixes
+
+- *Soulbinder's Nethermantle is no longer group looted.*
+- *Loombeast Silk and Ethereal Essence Silver (MFO transmog currency) are now ignored, and existing history entries removed.*
+
 # 3.17.1
 
 ## Bugfixes