+++ changelog.md
@@ -1,5 +1,11 @@
 ## Bugfixes
 
+- *Fixed broken Auto Award.*
+
+# 3.17.2
+
+## Bugfixes
+
 - *Soulbinder's Nethermantle is no longer group looted.*
 - *Loombeast Silk and Ethereal Essence Silver (MFO transmog currency) are now ignored, and existing history entries removed.*